+++ CHANGELOG.md
@@ -4,3 +4,7 @@
  - change version
  - create CHANGELOG.md
  - modify README.md
+
+ 05.01.2020 - [Benny Stark](https://github.com/Diggen85)
+ - create (auto)build environment for dev
+ - incomplete overall install!!! - raspion.deb builds